Dockerize build
There is a Dockerfile for an ubuntu base image.
There is also a NixDockerfile for a nixos base image. This one takes a
while to build from scratch, so the ubuntu base image is more practical.
Nixos also does not provide a native and clear way to pin specific
user-decided python package versions. Nixos channels are meant to supply
a specific version for each python package, however, specifying a
specific, unchanging, stable nix-channel was not straightforward with a
reasonable build-from-scratch build time.